# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BackupBudgetRow": {
        "description": "One budgets row (at most one per user — the PK is ``user_id``).",
        "properties": {
          "cats": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Cats",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "total",
          "cats",
          "updated_at"
        ],
        "title": "BackupBudgetRow",
        "type": "object"
      },
      "BackupCounts": {
        "description": "Row counts, mirroring the collections of the same envelope.",
        "properties": {
          "budgets": {
            "title": "Budgets",
            "type": "integer"
          },
          "debts": {
            "title": "Debts",
            "type": "integer"
          },
          "expenses": {
            "title": "Expenses",
            "type": "integer"
          },
          "recurring": {
            "default": 0,
            "title": "Recurring",
            "type": "integer"
          }
        },
        "required": [
          "expenses",
          "debts",
          "budgets"
        ],
        "title": "BackupCounts",
        "type": "object"
      },
      "BackupEnvelope": {
        "description": "The full-fidelity backup document (both directions, ADR-0012/0028).\n\n``expenses``/``debts`` reuse :class:`ExpenseOut`/:class:`DebtOut`, so the\nbackup wire shape IS the CRUD wire shape (every column, money as exact\ndecimal strings) — a downloaded file can be fed straight back to\n``POST /import/restore``. ``recurring`` (v2, ADR-0028) carries the\ncaller's rules; it defaults to empty so v1 consumers ignore it safely.",
        "properties": {
          "budgets": {
            "items": {
              "$ref": "#/components/schemas/BackupBudgetRow"
            },
            "title": "Budgets",
            "type": "array"
          },
          "counts": {
            "$ref": "#/components/schemas/BackupCounts"
          },
          "debts": {
            "items": {
              "$ref": "#/components/schemas/DebtOut"
            },
            "title": "Debts",
            "type": "array"
          },
          "expenses": {
            "items": {
              "$ref": "#/components/schemas/ExpenseOut"
            },
            "title": "Expenses",
            "type": "array"
          },
          "exported_at": {
            "title": "Exported At",
            "type": "string"
          },
          "recurring": {
            "items": {
              "$ref": "#/components/schemas/BackupRecurringRow"
            },
            "title": "Recurring",
            "type": "array"
          },
          "schema_version": {
            "const": 2,
            "default": 2,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "exported_at",
          "counts",
          "expenses",
          "debts",
          "budgets"
        ],
        "title": "BackupEnvelope",
        "type": "object"
      },
      "BackupRecurringRow": {
        "description": "One recurring-rule row (ADR-0028) — every column EXCEPT ``id``/``user_id``.\n\nRules have no history to preserve beyond their content, so unlike\nexpenses/debts the export omits identity entirely: restore always mints\nfresh PKs under the caller. Field validators/amount types are the shared\n:mod:`app.schemas.recurring` ones (do not drift — ADR-0004).",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "amt": {
            "title": "Amt",
            "type": "string"
          },
          "cat": {
            "title": "Cat",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "desc": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desc"
          },
          "freq": {
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly"
            ],
            "title": "Freq",
            "type": "string"
          },
          "grp": {
            "enum": [
              "food",
              "housing",
              "utility",
              "transport",
              "health",
              "education",
              "personal",
              "other"
            ],
            "title": "Grp",
            "type": "string"
          },
          "next_run": {
            "format": "date",
            "title": "Next Run",
            "type": "string"
          },
          "pay": {
            "enum": [
              "cash",
              "bkash",
              "nagad",
              "rocket",
              "card",
              "bank"
            ],
            "title": "Pay",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "cat",
          "grp",
          "amt",
          "pay",
          "freq",
          "start_date",
          "next_run",
          "active"
        ],
        "title": "BackupRecurringRow",
        "type": "object"
      },
...
      "DebtOut": {
        "description": "Public debt row; keys mirror the DB columns exactly.",
        "properties": {
          "amt": {
            "title": "Amt",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "dir": {
            "enum": [
              "lend",
              "borrow"
            ],
            "title": "Dir",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "iso": {
            "format": "date",
            "title": "Iso",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "party": {
            "title": "Party",
            "type": "string"
          },
          "settled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settled At"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "party",
          "dir",
          "amt",
          "iso",
          "created_at"
        ],
        "title": "DebtOut",
        "type": "object"
      },
...
      "ExpenseOut": {
        "description": "Public expense row; keys mirror the DB columns exactly.",
        "properties": {
          "amt": {
            "title": "Amt",
            "type": "string"
          },
          "cat": {
            "title": "Cat",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "desc": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desc"
          },
          "grp": {
            "enum": [
              "food",
              "housing",
              "utility",
              "transport",
              "health",
              "education",
              "personal",
              "other"
            ],
            "title": "Grp",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "iso": {
            "format": "date",
            "title": "Iso",
            "type": "string"
          },
          "pay": {
            "enum": [
              "cash",
              "bkash",
              "nagad",
              "rocket",
              "card",
              "bank"
            ],
            "title": "Pay",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "cat",
          "grp",
          "amt",
          "pay",
          "iso",
          "created_at"
        ],
        "title": "ExpenseOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RestoreOut": {
        "description": "POST /import/restore response.",
        "properties": {
          "restored": {
            "$ref": "#/components/schemas/BackupCounts"
          }
        },
        "required": [
          "restored"
        ],
        "title": "RestoreOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Poi Poi Hisab API",
    "version": "0.28.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/export/backup.json": {
      "get": {
        "description": "Dump the caller's complete ledger as one JSON envelope (ADR-0012).\n\nEvery column of every expenses/debts/budgets row; rows are ordered\ndeterministically (``iso ASC, id ASC`` — the CSV export's order) so the\nsame data always yields the same document modulo ``exported_at``.",
        "operationId": "export_backup_json_api_v1_export_backup_json_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupEnvelope"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Backup Json",
        "tags": [
          "export"
        ]
      }
    },
...
    "/api/v1/import/restore": {
      "post": {
        "description": "REPLACE the caller's ledger with the uploaded backup (v1/v2, ADR-0012).\n\nThe whole swap is one transaction: delete the caller's budgets, debts,\nexpenses AND recurring rules, insert the uploaded rows with fresh PKs,\ncommit. Recurring rows are wiped for BOTH versions — a v1 file predates\nrules, so restoring it means the ledger equals the file (ADR-0028).\nRows whose validation fails never reach the database; a failure\nmid-insert rolls the deletes back too.",
        "operationId": "import_restore_api_v1_import_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "description": "A GET /export/backup.json envelope (schema_version 1 or 2).",
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import Restore",
        "tags": [
          "import"
        ]
      }
    },
...
  }
}